requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.0
aioredis==2.0.1
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.3.4
certifi==2023.7.22
Django==4.2.7
django-cors-headers==4.7.0
django-modeltranslation==0.18.11
djangorestframework==3.15.2
frozenlist==1.7.0
idna==3.10
kombu==5.5.4
magic-filter==1.0.12
multidict==6.4.4
Pillow==10.1.0
propcache==0.3.2
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
yarl==1.20.1
